Move FormatDate Into a Component
- We need to format the scheduled for date in the SocialMediaPostDetail
  and SocialMediaPostTable.  I've refactored FormatDate into a
component that can be used for both.

- Improved the design a bit.
